requirements.txt
python-dotenv
streamlit
langchain
langchain_openai
langchain_community